add vdf | correction of git_commands.txt

+++ blkchn_code_/src/HashFunctions/SHA256/sha256.h
@@ -5,6 +5,6 @@
 #ifndef HASHFUNCTIONS_SHA256_H
 #define HASHFUNCTIONS_SHA256_H
 
-int sha256_fun(uint8_t data[], unsigned char *hash, uint8_t rounds, size_t data_len);
+int sha256_fun(uint8_t data[], unsigned char *hash, int rounds, size_t data_len);
 
 #endif //HASHFUNCTIONS_SHA256_H
